requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
black==22.3.0
click==8.1.3
Django~=4.0.4
django-environ==0.8.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-memcached==1.59
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
factory_boy

Faker~=13.14.0